# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Furtalk 评论系统的 HTTP API：覆盖站点与管理员引导、\n第一方与 widget 认证/会话、评论发布与审核、站点/用户/设置/\n提供商管理与邮件通知退订。",
        "title": "Furtalk API",
        "contact": {},
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/admin/sites": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-sites"
                ],
                "summary": "列出站点",
                "responses": {
                    "200": {
                        "description": "站点列表",
                        "schema": {
                            "$ref": "#/definitions/handler.SiteListResponse"
                        }
                    },
                    "401": {
                        "description": "需要管理员登录",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-sites"
                ],
                "summary": "创建站点",
                "parameters": [
                    {
                        "description": "站点名称与规范 URL",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.SiteRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "CSRF token",
                        "name": "X-CSRF-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "创建的站点",
                        "schema": {
                            "$ref": "#/definitions/handler.SiteResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数无效",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "需要管理员登录",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "站点冲突",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/sites/{site_id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-sites"
                ],
                "summary": "获取站点详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "站点 ID（十进制字符串）",
                        "name": "site_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "站点详情",
                        "schema": {
                            "$ref": "#/definitions/handler.SiteResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数无效",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "需要管理员登录",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "站点不存在",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "admin-sites"
                ],
                "summary": "删除站点",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "站点 ID（十进制字符串）",
                        "name": "site_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "删除需要显式确认",
                        "name": "confirm",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "CSRF token",
                        "name": "X-CSRF-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "删除完成"
                    },
                    "400": {
                        "description": "请求参数无效",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "需要管理员登录",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "站点不存在",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "资源状态冲突",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "破坏性操作需要显式确认",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-sites"
                ],
                "summary": "更新站点",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "站点 ID（十进制字符串）",
                        "name": "site_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "可更新的字段",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.SiteUpdateRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "CSRF token",
                        "name": "X-CSRF-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新后的站点",
                        "schema": {
                            "$ref": "#/definitions/handler.SiteResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数无效",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "需要管理员登录",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "站点不存在",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/sites/{site_id}/origins": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-sites"
                ],
                "summary": "添加站点允许的 origin",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "站点 ID（十进制字符串）",
                        "name": "site_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "精确 origin",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.OriginRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "CSRF token",
                        "name": "X-CSRF-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "创建的 origin",
                        "schema": {
                            "$ref": "#/definitions/handler.OriginResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数无效",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "需要管理员登录",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "站点不存在",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "资源状态冲突",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "origin 格式非法",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/sites/{site_id}/origins/{origin_id}": {
            "delete": {
                "tags": [
                    "admin-sites"
                ],
                "summary": "移除站点允许的 origin",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "站点 ID（十进制字符串）",
                        "name": "site_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "origin ID（十进制字符串）",
                        "name": "origin_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "CSRF token",
                        "name": "X-CSRF-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "移除完成"
                    },
                    "400": {
                        "description": "请求参数无效",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "需要管理员登录",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "站点或 origin 不存在",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-sites"
                ],
                "summary": "更新站点允许的 origin",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "站点 ID（十进制字符串）",
                        "name": "site_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "origin ID（十进制字符串）",
                        "name": "origin_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "精确 origin",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.OriginRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "CSRF token",
                        "name": "X-CSRF-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新后的 origin",
                        "schema": {
                            "$ref": "#/definitions/handler.OriginResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数无效",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "需要管理员登录",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "站点或 origin 不存在",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "资源状态冲突",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "origin 格式非法",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/me/comments/sites": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "me"
                ],
                "summary": "列出本人评论的站点选项",
                "responses": {
                    "200": {
                        "description": "本人发表过评论的站点",
                        "schema": {
                            "$ref": "#/definitions/handler.MeCommentSitesResponse"
                        }
                    },
                    "401": {
                        "description": "需要登录",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.MeCommentSiteResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "handler.MeCommentSitesResponse": {
            "type": "object",
            "properties": {
                "sites": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.MeCommentSiteResponse"
                    }
                }
            }
        },
...
        "handler.OriginRequest": {
            "type": "object",
            "properties": {
                "origin": {
                    "type": "string"
                }
            }
        },
        "handler.OriginResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "origin": {
                    "type": "string"
                }
            }
        },
...
        "handler.SiteListResponse": {
            "type": "object",
            "properties": {
                "sites": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.SiteResponse"
                    }
                }
            }
        },
        "handler.SiteRequest": {
            "type": "object",
            "properties": {
                "canonical_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "handler.SiteResponse": {
            "type": "object",
            "properties": {
                "canonical_url": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "origins": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.OriginResponse"
                    }
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "handler.SiteUpdateRequest": {
            "type": "object",
            "properties": {
                "canonical_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "httpx.ErrorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
        "httpx.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/httpx.ErrorBody"
                }
            }
        },
...
    }
}